Finally, something that can actually parse this.
Ok, so I had a brainwave - what if we put markers at where we need to execute some code - then we  can use those flags as markers for code to be executed, making it a lot easier for my tiny brain to do loops - therefore this is the edit where functionality can be added - then it's a lot easier. First real progress since ever. Part 1.